requirements.txt
Django==3.0.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
mysqlclient==2.0.3